+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 #### Unreleased
 
+- Allow field definitions to be defined as any `iterable`, not just `array`
+
 #### 14.10.0
 
 Feat: